feat:トレイト使用
